app/requirements.txt
# model run requirements
torch>=1.0.0
torchvision
dominate>=2.3.1
dill
numpy
scikit-image
# model serving requirements
spell
pillow==6.1
flask==1.1.1